Add custom Git subcommands
`git path` returns the current working directory or the path to the root
of the repository when running inside GitHub Actions or GitLab CI.

`git stem` returns the commit hash that likely is where the current
branch sprung out from. This hash is 100% certain when running inside
GitHub Actions or GitLab CI because those systems report it to us.